requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.14.0
keras==2.0.8
opencv-python
h5py==2.10.0
imgaug
IPython[all]
Shapely
torch
tqdm
torchvision
sklearn
pandas